atividade7

+++ arquivo 7.py
@@ -2,7 +2,7 @@
 n=int(input("digite quantos numeros"))
 soma=0
 for x in range(n):
-    valor=float(input("digite um valor"))
+    valor=int(input("digite um valor"))
     soma=soma+valor
 media=soma/n
